move config.js from /config to /server
